+++ braindecode/augmentation/transforms.py
@@ -3,19 +3,26 @@
 #
 # License: BSD (3-clause)
 
-from numbers import Real
 import warnings
+from numbers import Real
 
 import numpy as np
 import torch
 from mne.channels import make_standard_montage
 
 from .base import Transform
-from .functional import (
-    time_reverse, sign_flip, ft_surrogate, channels_dropout, channels_shuffle,
-    gaussian_noise, channels_permute, smooth_time_mask, bandstop_filter,
-    frequency_shift, sensors_rotation, mixup
-)
+from .functional import bandstop_filter
+from .functional import channels_dropout
+from .functional import channels_permute
+from .functional import channels_shuffle
+from .functional import frequency_shift
+from .functional import ft_surrogate
+from .functional import gaussian_noise
+from .functional import mixup
+from .functional import sensors_rotation
+from .functional import sign_flip
+from .functional import smooth_time_mask
+from .functional import time_reverse
 
 
 class TimeReverse(Transform):
@@ -109,7 +116,7 @@ class FTSurrogate(Transform):
             "phase_noise_magnitude should be between 0 and 1."
         self.phase_noise_magnitude = phase_noise_magnitude
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -121,12 +128,18 @@ class FTSurrogate(Transform):
 
         Returns
         -------
-        phase_noise_magnitude : float
-            The magnitude of the transformation.
-        rng : numpy.random.Generator
-            The generator to use.
+        params : dict
+            Contains:
+
+            * phase_noise_magnitude : float
+                The magnitude of the transformation.
+            * random_state : numpy.random.Generator
+                The generator to use.
         """
-        return self.phase_noise_magnitude, self.rng
+        return {
+            "phase_noise_magnitude": self.phase_noise_magnitude,
+            "random_state": self.rng,
+        }
 
 
 class ChannelsDropout(Transform):
@@ -166,7 +179,7 @@ class ChannelsDropout(Transform):
         )
         self.p_drop = p_drop
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -178,13 +191,19 @@ class ChannelsDropout(Transform):
 
         Returns
         -------
-        p_drop : float
-            Float between 0 and 1 setting the probability of dropping each
-            channel.
-        rng : numpy.random.Generator
-            The generator to use.
+        params : dict
+            Contains
+
+            * p_drop : float
+                Float between 0 and 1 setting the probability of dropping each
+                channel.
+            * random_state : numpy.random.Generator
+                The generator to use.
         """
-        return self.p_drop, self.rng
+        return {
+            "p_drop": self.p_drop,
+            "random_state": self.rng,
+        }
 
 
 class ChannelsShuffle(Transform):
@@ -198,7 +217,7 @@ class ChannelsShuffle(Transform):
         Float setting the probability of applying the operation.
     p_shuffle: float | None, optional
         Float between 0 and 1 setting the probability of including the channel
-        in the set of permutted channels. Defaults to 0.2.
+        in the set of permuted channels. Defaults to 0.2.
     random_state: int | numpy.random.Generator, optional
         Seed to be used to instantiate numpy random number generator instance.
         Used to decide whether or not to transform given the probability
@@ -225,7 +244,7 @@ class ChannelsShuffle(Transform):
         )
         self.p_shuffle = p_shuffle
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -237,13 +256,19 @@ class ChannelsShuffle(Transform):
 
         Returns
         -------
-        p_shuffle : float
-            Float between 0 and 1 setting the probability of including the
-            channel in the set of permutted channels.
-        rng : numpy.random.Generator
-            The generator to use.
+        params : dict
+            Contains
+
+            * p_shuffle : float
+                Float between 0 and 1 setting the probability of including the
+                channel in the set of permuted channels.
+            * random_state : numpy.random.Generator
+                The generator to use.
         """
-        return self.p_shuffle, self.rng
+        return {
+            "p_shuffle": self.p_shuffle,
+            "random_state": self.rng,
+        }
 
 
 class GaussianNoise(Transform):
@@ -288,7 +313,7 @@ class GaussianNoise(Transform):
         )
         self.std = std
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -300,12 +325,18 @@ class GaussianNoise(Transform):
 
         Returns
         -------
-        std : float
-            Standard deviation to use for the additive noise.
-        rng : numpy.random.Generator
-            The generator to use.
+        params : dict
+            Contains
+
+            * std : float
+                Standard deviation to use for the additive noise.
+            * random_state : numpy.random.Generator
+                The generator to use.
         """
-        return self.std, self.rng
+        return {
+            "std": self.std,
+            "random_state": self.rng,
+        }
 
 
 class ChannelsSymmetry(Transform):
@@ -367,7 +398,7 @@ class ChannelsSymmetry(Transform):
             permutation.append(new_position)
         self.permutation = permutation
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -379,10 +410,13 @@ class ChannelsSymmetry(Transform):
 
         Returns
         -------
-        permutation : float
-            List of integers defining the new channels order.
+        params : dict
+            Contains
+
+            * permutation : float
+                List of integers defining the new channels order.
         """
-        return (self.permutation,)
+        return {"permutation": self.permutation}
 
 
 class SmoothTimeMask(Transform):
@@ -430,7 +464,7 @@ class SmoothTimeMask(Transform):
         ), "mask_len_samples has to be a positive integer"
         self.mask_len_samples = mask_len_samples
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -442,14 +476,21 @@ class SmoothTimeMask(Transform):
 
         Returns
         -------
-        mask_start_per_sample : torch.tensor
-            Tensor of integers containing the position (in last dimension)
-            where to start masking the signal. Should have the same size as the
-            first dimension of X (i.e. one start position per example in the
-            batch).
-        mask_len_samples : int
-            Number of consecutive samples to zero out.
+        params : dict
+            Contains two elements:
+
+            * mask_start_per_sample : torch.tensor
+                Tensor of integers containing the position (in last dimension)
+                where to start masking the signal. Should have the same size as
+                the first dimension of X (i.e. one start position per example
+                in the batch).
+            * mask_len_samples : int
+                Number of consecutive samples to zero out.
         """
+        if len(batch) == 0:
+            return super().get_params(*batch)
+        X = batch[0]
+
         seq_length = torch.as_tensor(X.shape[-1], device=X.device)
         mask_len_samples = self.mask_len_samples
         if isinstance(mask_len_samples, torch.Tensor):
@@ -457,7 +498,10 @@ class SmoothTimeMask(Transform):
         mask_start = torch.as_tensor(self.rng.uniform(
             low=0, high=1, size=X.shape[0],
         ), device=X.device) * (seq_length - mask_len_samples)
-        return mask_start, mask_len_samples
+        return {
+            "mask_start_per_sample": mask_start,
+            "mask_len_samples": mask_len_samples,
+        }
 
 
 class BandstopFilter(Transform):
@@ -470,9 +514,9 @@ class BandstopFilter(Transform):
     ----------
     probability : float
         Float setting the probability of applying the operation.
-    bandwidth : float, optional
+    bandwidth : float
         Bandwidth of the filter, i.e. distance between the low and high cut
-        frequencies. Defaults to 1Hz.
+        frequencies.
     sfreq : float, optional
         Sampling frequency of the signals to be filtered. Defaults to 100 Hz.
     max_freq : float | None, optional
@@ -530,7 +574,7 @@ class BandstopFilter(Transform):
         self.max_freq = max_freq
         self.bandwidth = bandwidth
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -542,25 +586,37 @@ class BandstopFilter(Transform):
 
         Returns
         -------
-        sfreq : float
-            Sampling frequency of the signals to be filtered.
-        bandwidth : float
-            Bandwidth of the filter, i.e. distance between the low and high cut
-            frequencies.
-        freqs_to_notch : array-like | None
-            Array of floats of size ``(batch_size,)`` containing the center of
-            the frequency band to filter out for each sample in the batch.
-            Frequencies should be greater than ``bandwidth/2 + transition`` and
-            lower than ``sfreq/2 - bandwidth/2 - transition`` (where
-            ``transition = 1 Hz``).
+        params : dict
+            Contains
+
+            * sfreq : float
+                Sampling frequency of the signals to be filtered.
+            * bandwidth : float
+                Bandwidth of the filter, i.e. distance between the low and high
+                cut frequencies.
+            * freqs_to_notch : array-like | None
+                Array of floats of size ``(batch_size,)`` containing the center
+                of the frequency band to filter out for each sample in the
+                batch. Frequencies should be greater than
+                ``bandwidth/2 + transition`` and lower than
+                ``sfreq/2 - bandwidth/2 - transition`` (where
+                ``transition = 1 Hz``).
         """
+        if len(batch) == 0:
+            return super().get_params(*batch)
+        X = batch[0]
+
         # Prevents transitions from going below 0 and above max_freq
         notched_freqs = self.rng.uniform(
             low=1 + 2 * self.bandwidth,
             high=self.max_freq - 1 - 2 * self.bandwidth,
             size=X.shape[0]
         )
-        return self.sfreq, self.bandwidth, notched_freqs
+        return {
+            "sfreq": self.sfreq,
+            "bandwidth": self.bandwidth,
+            "freqs_to_notch": notched_freqs,
+        }
 
 
 class FrequencyShift(Transform):
@@ -600,7 +656,7 @@ class FrequencyShift(Transform):
 
         self.max_delta_freq = max_delta_freq
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -612,11 +668,18 @@ class FrequencyShift(Transform):
 
         Returns
         -------
-        delta_freq : float
-            The amplitude of the frequency shift (in Hz).
-        sfreq : float
-            Sampling frequency of the signals to be transformed.
+        params : dict
+            Contains
+
+            * delta_freq : float
+                The amplitude of the frequency shift (in Hz).
+            * sfreq : float
+                Sampling frequency of the signals to be transformed.
         """
+        if len(batch) == 0:
+            return super().get_params(*batch)
+        X = batch[0]
+
         u = torch.as_tensor(
             self.rng.uniform(size=X.shape[0]),
             device=X.device
@@ -625,7 +688,10 @@ class FrequencyShift(Transform):
         if isinstance(max_delta_freq, torch.Tensor):
             max_delta_freq = max_delta_freq.to(X.device)
         delta_freq = u * 2 * max_delta_freq - max_delta_freq
-        return delta_freq, self.sfreq
+        return {
+            "delta_freq": delta_freq,
+            "sfreq": self.sfreq,
+        }
 
 
 def _get_standard_10_20_positions(raw_or_epoch=None, ordered_ch_names=None):
@@ -731,7 +797,7 @@ class SensorsRotation(Transform):
         self.spherical_splines = spherical_splines
         self.max_degrees = max_degrees
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -743,21 +809,29 @@ class SensorsRotation(Transform):
 
         Returns
         -------
-        sensors_positions_matrix : numpy.ndarray
-            Matrix giving the positions of each sensor in a 3D cartesian
-            coordinate system. Should have shape (3, n_channels), where
-            n_channels is the number of channels.
-        axis : 'x' | 'y' | 'z'
-            Axis around which to rotate.
-        angles : array-like
-            Array of float of shape ``(batch_size,)`` containing the rotation
-            angles (in degrees) for each element of the input batch, sampled
-            uniformly between ``-max_degrees``and ``max_degrees``.
-        spherical_splines : bool
-            Whether to use spherical splines for the interpolation or not. When
-            `False`, standard scipy.interpolate.Rbf (with quadratic kernel)
-            will be used (as in the original paper).
+        params : dict
+            Contains four elements:
+
+            * sensors_positions_matrix : numpy.ndarray
+                Matrix giving the positions of each sensor in a 3D cartesian
+                coordinate system. Should have shape (3, n_channels), where
+                n_channels is the number of channels.
+            * axis : 'x' | 'y' | 'z'
+                Axis around which to rotate.
+            * angles : array-like
+                Array of float of shape ``(batch_size,)`` containing the
+                rotation angles (in degrees) for each element of the input
+                batch, sampled uniformly between ``-max_degrees``and
+                ``max_degrees``.
+            * spherical_splines : bool
+                Whether to use spherical splines for the interpolation or not.
+                When ``False``, standard scipy.interpolate.Rbf (with quadratic
+                kernel) will be used (as in the original paper).
         """
+        if len(batch) == 0:
+            return super().get_params(*batch)
+        X = batch[0]
+
         u = self.rng.uniform(
             low=0,
             high=1,
@@ -768,12 +842,12 @@ class SensorsRotation(Transform):
             max_degrees = max_degrees.to(X.device)
         random_angles = torch.as_tensor(
             u, device=X.device) * 2 * max_degrees - max_degrees
-        return (
-            self.sensors_positions_matrix,
-            self.axis,
-            random_angles,
-            self.spherical_splines
-        )
+        return {
+            "sensors_positions_matrix": self.sensors_positions_matrix,
+            "axis": self.axis,
+            "angles": random_angles,
+            "spherical_splines": self.spherical_splines
+        }
 
 
 class SensorsZRotation(SensorsRotation):
@@ -982,7 +1056,7 @@ class Mixup(Transform):
         self.alpha = alpha
         self.beta_per_sample = beta_per_sample
 
-    def get_params(self, X, y):
+    def get_params(self, *batch):
         """Return transform parameters.
 
         Parameters
@@ -994,12 +1068,13 @@ class Mixup(Transform):
 
         Returns
         -------
-        lam : torch.Tensor
-            Values sampled uniformly between 0 and 1 setting the linear
-            interpolation between examples.
-        idx_perm: torch.Tensor
-            Shuffled indices of example that are mixed into original examples.
+        params: dict
+            Contains the values sampled uniformly between 0 and 1 setting the
+            linear interpolation between examples (lam) and the shuffled
+            indices of examples that are mixed into original examples
+            (idx_perm).
         """
+        X = batch[0]
         device = X.device
         batch_size, _, _ = X.shape
 
@@ -1016,4 +1091,7 @@ class Mixup(Transform):
 
         idx_perm = torch.as_tensor(self.rng.permutation(batch_size,))
 
-        return lam, idx_perm
+        return {
+            "lam": lam,
+            "idx_perm": idx_perm,
+        }