+++ docs/whats_new.rst
@@ -40,6 +40,7 @@ Enhancements
 - Adding time series targets decoding together with :class:`braindecode.datasets.BCICompetitionIVDataset4` and fingers flexion decoding from ECoG examples (:gh:`261` by `Maciej Śliwowski`_ and `Mohammed Fattouh`_)
 - Make EEGClassifier and EEGRegressor cloneable for scikit-learn :class:`braindecode.augmentation.Mixup` (:gh:`347` by `Lukas Gemein`_, `Robin Tibor Schirrmeister`_, `Maciej Śliwowski`_ and `Alex Gramfort`_)
 - Allow to raise a warning when a few trials are shorter than the windows length, instead of raising an error and stopping all computation. (:gh:`353` by `Cedric Rommel`_)
+- Setting `torch.backends.cudnn.benchmark` in :func:`braindecode.util.set_random_seeds`, adding warning and more info to the docstring to imporve reproducibility (:gh:`333` by `Maciej Śliwowski`_)
 
 Bugs
 ~~~~